requirements.txt
mysql-connector-python
numpy
tqdm

